parking_finder/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
Fiona==1.9.4.post1
geopandas==0.13.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
packaging==23.1
pandas==2.0.3
postgis==1.0.4
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
